+++ geo-types/CHANGES.md
@@ -1,5 +1,7 @@
 # Changes
 
+## unreleased
+
 ## 0.7.8
 
 * Rename `Coordinate` to `Coord`; add deprecated `Coordinate` that is an alias for `Coord`